Reorganize and implement categoryDto
