+++ src/main/java/com/laboration2/category/CategoryDto.java
@@ -1,6 +1,6 @@
-package com.laboration2.dto;
+package com.laboration2.category;
 
-import com.laboration2.entities.Category;
+import com.laboration2.location.Location;
 import jakarta.validation.constraints.Size;
 
 import java.io.Serializable;
@@ -10,5 +10,5 @@ import java.util.Set;
  * DTO for {@link Category}
  */
 public record CategoryDto(Long id, @Size(max = 255) String name, @Size(max = 255) String symbol,
-                          @Size(max = 255) String description, Set<LocationDto> locations) implements Serializable {
+                          @Size(max = 255) String description, Set<Location> locations) implements Serializable {
 }
