+++ src/main/java/com/laboration2/category/CategoryRepository.java
@@ -1,0 +1,10 @@
+package com.laboration2.category;
+
+import com.laboration2.category.Category;
+import org.springframework.data.repository.ListCrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CategoryRepository extends ListCrudRepository<Category, Long> {
+
+}
