+++ src/main/java/com/laboration2/location/LocationDto.java
@@ -1,6 +1,7 @@
-package com.laboration2.dto;
+package com.laboration2.location;
 
-import com.laboration2.entities.Location;
+import com.laboration2.category.CategoryDto;
+import com.laboration2.user.UserDto;
 import jakarta.validation.constraints.Size;
 
 import java.io.Serializable;
